Add files via upload
it is an object-oriented programming with C++. in this file i have used to make an structure for array which called myarray. in this code i have used structure instead of class, because they have a different accessibility(Private or public). myarray has two properties: the first one is 'n' which store the size of array and the second one is '*a' which store the address of first home of array. constructor method , copy constructor method , deconstructor method and some other methods like add, sub, mul, neg, print.